requirements.txt
discord
aiohttp
py-hPickle